backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.7.8
pyodbc==5.0.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
azure-identity==1.15.0
azure-storage-blob==12.19.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
